Update(boundary conditions): Allow tolerance when defining WMESH
Because of roundoff errors, sum(WMESH), defined in boundary conditions, might not be necessarily exactly equal to LENGTH (defined in the geometry file). A tolerance of 1 mm is allowed in BoundaryConditions.m/BoundaryConditions.
Because of potential issue when extrapolating the axial power shape within that tolerance (for instance using the ’next’ method), any NaN in the extrapolated power is re-calculated using a linear assumption in MixtureSolver.m/axialInterpolate.