fix(fourfield): initial implementation of fourfield solver
Due to changes in implementation of W, U, and L in FF, an AbstractFilm
method was created to encapsulate most TF.Film methods while allowing
flexiblity on certain property implementation.

Currently, FF does not converge on cases validated by the TF solver even
though the former is implemented to behave identically as the latter. Further
investigation is needed.