refactor: address issue #52
Utility functions `struct` and `copyFlowProperties` in all `Solvers.AbstractField` subclasses are moved to `AbstractField`. Due to slight differences in the specific flow properties in each subclass, a new setAccess protected property is created in `AbstractField`: `flowProperties`. The two functions use the entries in this property to determine which properties to convert/copy. To overwrite the preset (`W, U, H`), a commented line is included in the constructor for each subclass that was modified as part of this update. See `Solvers.Mixture.Mixture` for how this overwriting is done.