fix(iss #54): implemented REGIMES enum class
The call to the built-in function `categorical` was the bottleneck. By
replacing `categorical` to a enum class construction, the same code
readability is achieved while vastly increasing execution time. This is
particularly the case since type `double` was specifically inherited by
the enum class, so theoretically less type-casting is needed behind the
scenes. However this resulting performance from this last claim was not explicitly tested

Nearly 5x faster computational speed achieved. Closing.