feat:Generic Okawa Model
This push is the most recent iteration of the Okawa entrainment model feature. Using any OKAWA model, a private method OKAWAMENT will be called to calculate the entrainment mass flux using a generalized equation with three coefficients, set depending on the model chosen by the user. This method can handle the 2003 and 2004 model formulations, or a user defined array of coefficients and thresholds. The format for the single array provided by the user is [a0 b0 c0 t1 a1 b1 c1 ...] where a, b, and c are the model coefficients and t are the model thresholds.

Addresses: #57