feat(mixture solver): add Groeneveld LUT for CBT
This implements an option for the user to select the 2006 Groeneveld LUT for
calculating where in the channel critical heat flux occurs. It includes K1
and K4 factors. Future updates may address storing the LUT data locally
so that is need not be read into memory each time the function is
called.